Add tests for safe_mode
